# docs/swagger.yaml
basePath: /api/v1
definitions:
  food.FoodList:
    properties:
      food_list:
        items:
          $ref: '#/definitions/service.FoodDetailsForRecommend'
        type: array
    type: object
  model.Menu:
    properties:
      name:
        type: string
      price:
        type: number
    type: object
  model.RestaurantDetails:
    properties:
      average_price:
        type: number
      introduction:
        type: string
      menus:
        items:
          $ref: '#/definitions/model.Menu'
        type: array
      name:
        type: string
      picture_url:
        type: string
    type: object
  restaurant.ResaurantListResponse:
    properties:
      restaurants:
        items:
          $ref: '#/definitions/service.RestaurantForCanteen'
        type: array
    type: object
  search.SearchFoodResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/service.SearchFoodModel'
        type: array
    type: object
  search.SearchRestaurantResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/service.SearchRestaurantModel'
        type: array
    type: object
  service.FoodDetailsForRecommend:
    properties:
      canteen_name:
        type: string
      ingredient:
        type: string
      introduction:
        type: string
      name:
        type: string
      picture_url:
        type: string
      resaurant_name:
        type: string
      storey:
        type: integer
    type: object
  service.RandomRestaurant:
    properties:
      average_price:
        type: number
      canteen_name:
        type: string
      picture_url:
        type: string
      recommendation:
        type: string
      restaurant_name:
        type: string
      storey:
        type: integer
    type: object
  service.RestaurantForCanteen:
    properties:
      average_price:
        type: number
      picture_url:
        type: string
      restaurant_id:
        type: integer
      restaurant_name:
        type: string
    type: object
  service.SearchFoodModel:
    properties:
      canteen_name:
        type: string
      name:
        type: string
      picture_url:
        type: string
      restaurant_name:
        type: string
      storey:
        type: integer
    type: object
  service.SearchRestaurantModel:
    properties:
      canteen_name:
        type: string
      name:
        type: string
      picture_url:
        type: string
      restaurant_id:
        type: integer
      storey:
        type: integer
    type: object
host: '....'
info:
  contact: {}
  description: 美食服务
  license: {}
  title: food_service
  version: "1.0"
paths:
  /food/recommend:
    get:
      consumes:
      - application/json
      description: 返回一些推荐的food
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页最大数
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/food.FoodList'
      summary: 华师必吃
      tags:
      - food
  /restaurant/detail/{id}:
    get:
      consumes:
      - application/json
      description: 给店家id，返回店家详情页
      parameters:
      - description: 店家的id，别的api会给出。
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.RestaurantDetails'
      summary: 店家详情页
      tags:
      - restaurant
  /restaurant/list:
    get:
      consumes:
      - application/json
      description: 返回一些推荐的食堂
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页最大数
        in: query
        name: limit
        required: true
        type: integer
      - description: 按照id来给,食堂信息，哪个食堂，楼层
        in: query
        name: c
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/restaurant.ResaurantListResponse'
      summary: 在线菜单
      tags:
      - restaurant
  /restaurant/random:
    get:
      consumes:
      - application/json
      description: 美食首页的推荐窗口
      parameters:
      - description: 食堂名字
        in: body
        name: canteen_name
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.RandomRestaurant'
      summary: 美食首页
      tags:
      - restaurant
  /search/food:
    get:
      consumes:
      - application/json
      description: 搜索返回一个list
      parameters:
      - description: 搜索信息, search_text
        in: query
        name: st
        required: true
        type: string
      - description: 第几页， page
        in: query
        name: page
        required: true
        type: integer
      - description: 每页多少个， limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/search.SearchFoodResponse'
      summary: 搜索食物
      tags:
      - search
  /search/restaurant:
    get:
      consumes:
      - application/json
      description: 搜索返回一个list
      parameters:
      - description: 搜索信息, search_text
        in: query
        name: st
        required: true
        type: string
      - description: 第几页， page
        in: query
        name: page
        required: true
        type: integer
      - description: 每页多少个， limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/search.SearchRestaurantResponse'
      summary: 搜索餐厅
      tags:
      - search
swagger: "2.0"
tags:
- description: 店铺(窗口)相关
  name: restaurant
- description: 菜品相关
  name: food
- description: 搜索相关
  name: search
- description: 食堂相关
  name: canteen